Rename contracts to Base*

+++ src/CryticAsserts.sol
@@ -1,52 +1,32 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import "./Asserts.sol";
+import {Asserts} from "./Asserts.sol";
 
 contract CryticAsserts is Asserts {
     event Log(string);
 
-    function gt(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function gt(uint256 a, uint256 b, string memory reason) internal virtual override {
         emit Log(reason);
         assert(a > b);
     }
 
-    function gte(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function gte(uint256 a, uint256 b, string memory reason) internal virtual override {
         emit Log(reason);
         assert(a >= b);
     }
 
-    function lt(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function lt(uint256 a, uint256 b, string memory reason) internal virtual override {
         emit Log(reason);
         assert(a < b);
     }
 
-    function lte(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function lte(uint256 a, uint256 b, string memory reason) internal virtual override {
         emit Log(reason);
         assert(a <= b);
     }
 
-    function eq(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function eq(uint256 a, uint256 b, string memory reason) internal virtual override {
         emit Log(reason);
         assert(a == b);
     }
@@ -56,28 +36,20 @@ contract CryticAsserts is Asserts {
         assert(b);
     }
 
-    function between(
-        uint256 value,
-        uint256 low,
-        uint256 high
-    ) internal virtual override returns (uint256) {
+    function between(uint256 value, uint256 low, uint256 high) internal virtual override returns (uint256) {
         if (value < low || value > high) {
-            uint ans = low + (value % (high - low + 1));
+            uint256 ans = low + (value % (high - low + 1));
             return ans;
         }
         return value;
     }
 
-    function between(
-        int256 value,
-        int256 low,
-        int256 high
-    ) internal virtual override returns (int256) {
+    function between(int256 value, int256 low, int256 high) internal virtual override returns (int256) {
         if (value < low || value > high) {
-            int range = high - low + 1;
-            int clamped = (value - low) % (range);
+            int256 range = high - low + 1;
+            int256 clamped = (value - low) % (range);
             if (clamped < 0) clamped += range;
-            int ans = low + clamped;
+            int256 ans = low + clamped;
             return ans;
         }
         return value;
