+++ src/FoundryAsserts.sol
@@ -1,47 +1,27 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import "forge-std/Test.sol";
-import "./Asserts.sol";
+import {Test} from "forge-std/Test.sol";
+import {Asserts} from "./Asserts.sol";
 
 contract FoundryAsserts is Test, Asserts {
-    function gt(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function gt(uint256 a, uint256 b, string memory reason) internal virtual override {
         assertGt(a, b, reason);
     }
 
-    function gte(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function gte(uint256 a, uint256 b, string memory reason) internal virtual override {
         assertGe(a, b, reason);
     }
 
-    function lt(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function lt(uint256 a, uint256 b, string memory reason) internal virtual override {
         assertLt(a, b, reason);
     }
 
-    function lte(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function lte(uint256 a, uint256 b, string memory reason) internal virtual override {
         assertLe(a, b, reason);
     }
 
-    function eq(
-        uint256 a,
-        uint256 b,
-        string memory reason
-    ) internal virtual override {
+    function eq(uint256 a, uint256 b, string memory reason) internal virtual override {
         assertEq(a, b, reason);
     }
 
@@ -49,36 +29,26 @@ contract FoundryAsserts is Test, Asserts {
         assertTrue(b, reason);
     }
 
-    function between(
-        uint256 value,
-        uint256 low,
-        uint256 high
-    ) internal virtual override returns (uint256) {
+    function between(uint256 value, uint256 low, uint256 high) internal virtual override returns (uint256) {
         if (value < low || value > high) {
-            uint ans = low + (value % (high - low + 1));
+            uint256 ans = low + (value % (high - low + 1));
             return ans;
         }
         return value;
     }
 
-    function between(
-        int256 value,
-        int256 low,
-        int256 high
-    ) internal virtual override returns (int256) {
+    function between(int256 value, int256 low, int256 high) internal virtual override returns (int256) {
         if (value < low || value > high) {
-            int range = high - low + 1;
-            int clamped = (value - low) % (range);
+            int256 range = high - low + 1;
+            int256 clamped = (value - low) % (range);
             if (clamped < 0) clamped += range;
-            int ans = low + clamped;
+            int256 ans = low + clamped;
             return ans;
         }
         return value;
     }
 
-    function precondition(
-        bool p
-    ) internal virtual override {
+    function precondition(bool p) internal virtual override {
         vm.assume(p);
     }
 }
